news-aggregator/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
pydantic==1.8.2
requests==2.26.0
beautifulsoup4==4.9.3
pika==1.2.0
aiohttp==3.8.1
newsapi-python==0.2.6
transformers==4.11.3
torch==1.9.0
sqlalchemy==1.4.23
aiosqlite==0.17.0
apscheduler==3.8.1
python-dotenv==0.19.0
tenacity==8.0.1
